Allow hidden items in the input-model
Allow the option of have hidden items in the elements array. I'm my case using the multiselect for hide/show columns in a table, the array with the columns definition have extra data needed by the table-plugin, with this extra parameter i can keep that extra data in the array without need to pass a filtered array to the multiselect and re-build the original array using the output-model data.