Making the label properties be templateable
Rather than just assuming each group of characters between spaces is a property of the model, I've made it iterate over every group of [\w] characters and all [^\w] characters. If a character group is a property on the model, then it'll append the value of that property in the model, otherwise it just output the group of characters.