Allows customization of icons
Fixes #245 in original repository.

After this change the user can fully customize the "icons", replacing
the default html entity characters by any string, including:

- Different html entities or unicode characters;
- Bits of html offloading things to a icon font;
- Html img tags;
- Empty strings (that effectively removes the icons);

The change is backwards compatible, only "activated" if the element that
uses this directive also has a `custom-icon` attribute. That attribute
works analogous to the `translation` attribute. Backwards compatability
is the reason for the somewhat convoluted "pad-with-nbsp" function, that
might be removed in a future major release.